Use BigInts instead of integers for i64

+++ migrations/2024-07-12-144642_create_users/up.sql
@@ -1,6 +1,6 @@
 -- Your SQL goes here
 CREATE TABLE users (
-    id INTEGER PRIMARY KEY NOT NULL,
-    points INTEGER NOT NULL,
+    id BIGINT PRIMARY KEY NOT NULL,
+    points BIGINT NOT NULL,
     username TEXT NOT NULL
 )
